+++ packages/orq-rc/RELEASES.md
@@ -7489,3 +7489,13 @@ Based on:
 - [typescript v4.5.0-rc.36] packages/orq-rc
 ### Releases
 - [NPM v4.5.0-rc.36] https://www.npmjs.com/package/@orq-ai/node/v/4.5.0-rc.36 - packages/orq-rc
+
+## 2026-03-06 12:48:28
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.745.0 (2.850.7) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.5.0-rc.37] packages/orq-rc
+### Releases
+- [NPM v4.5.0-rc.37] https://www.npmjs.com/package/@orq-ai/node/v/4.5.0-rc.37 - packages/orq-rc